requirements.txt
bilibili-api-python==17.2.0
fastapi==0.112.2
uvicorn[standard]==0.30.6
pydantic==2.9.0
pydantic-core==2.23.2
pywebview==5.3
Pillow==10.4.0
httpx==0.27.2
aiohttp==3.10.5
websockets==13.0
protobuf==6.31.1
pydub==0.25.1
python-multipart==0.0.20